+++ packages/search/__tests__/keyword.test.tsx
@@ -5,10 +5,11 @@ import { render } from '@testing-library/react';
 import { mockIsIntersecting } from '../../../test-utils/mockIntersectionObserver';
 import { Viewer } from '@react-pdf-viewer/core';
 import { searchPlugin } from '../src/index';
+import type { FlagKeyword } from '../src/types/FlagKeyword';
 
 const TestKeywordOption: React.FC<{
     fileUrl: Uint8Array;
-    keyword: string;
+    keyword: string | FlagKeyword;
 }> = ({ fileUrl, keyword }) => {
     const searchPluginInstance = searchPlugin({
         keyword,
@@ -43,5 +44,46 @@ test('keyword option', async () => {
     const highlights = await findAllByTitle(page, keyword);
     expect(highlights.length).toEqual(13);
     expect(highlights[0].getAttribute('title')).toEqual(keyword);
+    expect(highlights[0]).toHaveClass('rpv-search__highlight');
+});
+
+test('Special character in the keyword', async () => {
+    const keyword = '(a';
+
+    const { findByText, findByTestId, getByTestId } = render(
+        <TestKeywordOption fileUrl={global['__OPEN_PARAMS_PDF__']} keyword={keyword} />
+    );
+    mockIsIntersecting(getByTestId('core__viewer'), true);
+
+    const page = await findByTestId('core__page-layer-4');
+    mockIsIntersecting(page, true);
+
+    await findByText('Parameters for Opening PDF Files');
+
+    const highlights = await findAllByTitle(page, keyword);
+    expect(highlights.length).toEqual(1);
+    expect(highlights[0].getAttribute('title')).toEqual(keyword);
+    expect(highlights[0]).toHaveClass('rpv-search__highlight');
+});
+
+test('Match case of special character', async () => {
+    const flagKeyword = {
+        keyword: '(A',
+        matchCase: true,
+    };
+
+    const { findByText, findByTestId, getByTestId } = render(
+        <TestKeywordOption fileUrl={global['__OPEN_PARAMS_PDF__']} keyword={flagKeyword} />
+    );
+    mockIsIntersecting(getByTestId('core__viewer'), true);
+
+    const page = await findByTestId('core__page-layer-5');
+    mockIsIntersecting(page, true);
+
+    await findByText('Adobe Acrobat SDK');
+
+    const highlights = await findAllByTitle(page, flagKeyword.keyword);
+    expect(highlights.length).toEqual(1);
+    expect(highlights[0].getAttribute('title')).toEqual(flagKeyword.keyword);
     expect(highlights[0]).toHaveClass('rpv-search__highlight');
 });