+++ CHANGELOG.md
@@ -18,6 +18,7 @@ const defaultLayoutPluginInstance = defaultLayoutPlugin({
 
 **Bug fixes**
 
+-   Can't use special characters such as `(`, `)`, `[`, `]`, `*` in the keyword
 -   The highlight area aren't displayed
 -   The `selectedText` prop of `RenderHighlightContentProps` isn't correct